Moved tests into module
