+++ auth/auth-service-testing/src/main/java/net/squanchy/service/UserFixtures.kt
