+++ schedule/schedule-domain/src/test/java/net/squanchy/schedule/domain/view/ExperienceLevelTest.kt
@@ -1,4 +1,4 @@
-package net.squanchy.eventdetails.domain.view
+package net.squanchy.schedule.domain.view
 
 import com.google.common.truth.Truth.assertThat
 import net.squanchy.schedule.domain.view.ExperienceLevel
